requirements.txt
anthropic
datasets
google-generativeai
openai
Pillow
reka-api
scipy
tiktoken